move pre computed maps to the `signatures`
This will allow us to reuse them later in other migrations, once we
split the arbitrary-utilities migration to also handle the bare value to
named value migration.
